Fix: Correct Tailwind directive order in globals.css
Ensures that `globals.css` uses the standard `@tailwind base;`,
`@tailwind components;`, and `@tailwind utilities;` directives in the
recommended order. This is to address a build error where utility
classes used with `@apply` in imported stylesheets (like components.css)
were not being recognized.

The `components.css` file, which defines custom component classes
using `@apply`, is imported within the `@tailwind components;` layer context,
and other custom stylesheets are organized accordingly.
