+++ src/app/styles/components.css
@@ -30,4 +30,43 @@
            hover:bg-[var(--color-brand-blue-darker)]
            transition-colors;
   }
+
+  /* Booking Modal Styles */
+  .booking-dialog-panel {
+    @apply bg-white p-6 rounded-2xl shadow-xl w-full max-w-[42rem] relative; /* Added relative for close button positioning */
+  }
+
+  .dialog-title {
+    @apply text-[1.8rem] font-semibold text-gray-800 mb-4; /* Standardized title */
+  }
+
+  .close-button {
+    @apply absolute top-4 right-4 p-2 rounded-full text-gray-500 hover:text-gray-700 hover:bg-gray-100 transition-colors;
+  }
+
+  .selected-datetime {
+    @apply text-center text-gray-700 py-3 px-4 bg-gray-100 rounded-lg border border-gray-200;
+  }
+
+  .confirm-button {
+    @apply w-full mt-4 btn-primary py-3 text-base; /* py-3.5 in original btn-primary, adjusted to py-3 */
+  }
+
+  .btn-time-slot {
+    @apply p-3 rounded-lg border text-center transition-all h-12 flex items-center justify-center font-medium;
+  }
+  .btn-time-slot-selected {
+    @apply bg-[#7eaedb] text-white border-[#7eaedb] shadow-md;
+  }
+  .btn-time-slot-default {
+    @apply border-gray-300 hover:bg-[#5a82a3] hover:text-white hover:border-[#5a82a3];
+  }
+
+  .booking-step-title {
+    @apply flex items-center mb-3 text-lg font-medium text-gray-700;
+  }
+
+  .booking-step-number {
+    @apply flex items-center justify-center w-7 h-7 rounded-full bg-[#7eaedb] text-white mr-2 text-sm font-bold;
+  }
 }
