+++ src/app/styles/globals.css
@@ -1,10 +1,16 @@
-/* src/styles/globals.css */
-@import "tailwindcss";
-@import "react-datepicker/dist/react-datepicker.css";
+/* src/app/styles/globals.css */
+@tailwind base;
 
-/* Importer vores egne CSS-filer */
+/* Custom base styles and theme variables */
 @import "./theme.css";
 @import "./base.css";
-@import "./animations.css";
-@import "./components.css";
-@import "./datepicker.css";
+
+@tailwind components;
+
+/* Custom component classes using @apply */
+@import "./components.css"; 
+@import "./animations.css"; /* If it contains component-like classes */
+@import "react-datepicker/dist/react-datepicker.css"; /* Datepicker styles */
+@import "./datepicker.css"; /* Custom overrides for datepicker */
+
+@tailwind utilities;
